Time: 85 ms (42.57%), Space: 26.1 MB (68.10%) - LeetHub

+++ 1011-capacity-to-ship-packages-within-d-days/1011-capacity-to-ship-packages-within-d-days.cpp
@@ -5,7 +5,7 @@ public:
         int count = 0;
         int n = weights.size();
         int i = 0;
-        long long int w;
+        int w;
         while (i < n)
         {
             w = 0;
